Refactored graphics classes into new package
