+++ src/graphics/ControlPanel.java
@@ -1,3 +1,5 @@
+package graphics;
+
 import java.awt.Color;
 import java.awt.Dimension;
 import java.awt.event.ActionEvent;
@@ -9,6 +11,7 @@ import javax.swing.JButton;
 import javax.swing.JLabel;
 import javax.swing.JPanel;
 
+import main.Dice;
 import main.Player;
 import main.Settings;
