+++ src/graphics/PlayerPanel.java
@@ -1,3 +1,5 @@
+package graphics;
+
 import java.awt.BorderLayout;
 import java.awt.Color;
