+++ test/DiceTester.java
@@ -1,5 +1,8 @@
 import javax.swing.JFrame;
 
+import main.Dice;
+import graphics.DiceComponent;
+
 public class DiceTester {
     public static void main(String[] args) {
         JFrame frame = new JFrame();
@@ -7,7 +10,7 @@ public class DiceTester {
         frame.setTitle("Dice Tester");
         frame.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 
-        Dice dice = new Dice(50, 35);
+        Dice dice = new Dice(70, 50);
         DiceComponent diceImage = new DiceComponent(dice);
         dice.roll();
         System.out.println(dice.getSide());
